Add copyright and licensing information to files
This commit adds copyright notices and MIT License information to multiple source files in the project. The changes enhance legal clarity regarding the use and distribution of the code.

Namespace declarations in `DesignTimeDbContextFactory.cs`, `DeviceStatus.cs`, and `EmailSettings.cs` were updated to include the new copyright notice. The `HomeMonitoring.sln.DotSettings` file was modified to reflect the copyright in code style settings. Additionally, a new `LICENSE` file was added, outlining the terms of the MIT License.